+++ lib/src/cli/commands/history_command.dart
@@ -61,7 +61,10 @@ class HistoryCommand extends AuthenticatedCommand {
       final thread = message.replyCount != null
           ? ' [${message.replyCount} replies]'
           : '';
-      logger.info('[${message.ts}] <$user> ${message.text}$thread');
+      logger.info(
+        '[${message.ts}] <$user> ${message.text}'
+        '${message.attachmentSuffix}$thread',
+      );
     }
 
     if (page.hasMore) {